requirements.txt
streamlit
pandas
PyMuPDF
langchain
python-dotenv
openai
google-generativeai
langchain-google-genai